+++ src/widgets/splits/SplitOverlay.cpp
@@ -14,14 +14,117 @@
 #include <QPainter>
 #include <QPushButton>
 
+namespace {
+
+using namespace chatterino;
+
+class ButtonEventFilter : public QObject
+{
+    SplitOverlay *parent;
+    SplitOverlayButton buttonType;
+
+public:
+    ButtonEventFilter(SplitOverlay *_parent, SplitOverlayButton _buttonType);
+
+protected:
+    bool eventFilter(QObject *watched, QEvent *event) override;
+};
+
+ButtonEventFilter::ButtonEventFilter(SplitOverlay *_parent,
+                                     SplitOverlayButton _buttonType)
+    : QObject(_parent)
+    , parent(_parent)
+    , buttonType(_buttonType)
+{
+}
+
+bool ButtonEventFilter::eventFilter(QObject *watched, QEvent *event)
+{
+    switch (event->type())
+    {
+        case QEvent::Enter: {
+            auto *effect = dynamic_cast<QGraphicsOpacityEffect *>(
+                ((QWidget *)watched)->graphicsEffect());
+
+            if (effect != nullptr)
+            {
+                effect->setOpacity(0.99);
+            }
+
+            this->parent->setHoveredButton(this->buttonType);
+        }
+        break;
+
+        case QEvent::Leave: {
+            auto *effect = dynamic_cast<QGraphicsOpacityEffect *>(
+                ((QWidget *)watched)->graphicsEffect());
+
+            if (effect != nullptr)
+            {
+                effect->setOpacity(0.7);
+            }
+
+            this->parent->setHoveredButton({});
+        }
+        break;
+
+        case QEvent::MouseButtonPress: {
+            if (this->buttonType == SplitOverlayButton::Move)
+            {
+                auto *mouseEvent = dynamic_cast<QMouseEvent *>(event);
+                assert(mouseEvent != nullptr);
+                if (mouseEvent->button() == Qt::LeftButton)
+                {
+                    this->parent->dragPressed();
+                }
+                return true;
+            }
+        }
+        break;
+
+        case QEvent::MouseButtonRelease: {
+            switch (this->buttonType)
+            {
+                case SplitOverlayButton::Move:
+                    break;
+
+                case SplitOverlayButton::Left: {
+                    this->parent->createSplitPressed(SplitDirection::Left);
+                }
+                break;
+
+                case SplitOverlayButton::Up: {
+                    this->parent->createSplitPressed(SplitDirection::Above);
+                }
+                break;
+
+                case SplitOverlayButton::Right: {
+                    this->parent->createSplitPressed(SplitDirection::Right);
+                }
+                break;
+
+                case SplitOverlayButton::Down: {
+                    this->parent->createSplitPressed(SplitDirection::Below);
+                }
+                break;
+            }
+        }
+        break;
+
+        default:;
+    }
+    return QObject::eventFilter(watched, event);
+}
+
+}  // namespace
+
 namespace chatterino {
 
 SplitOverlay::SplitOverlay(Split *parent)
     : BaseWidget(parent)
     , split_(parent)
 {
-    QGridLayout *layout = new QGridLayout(this);
-    this->layout_ = layout;
+    auto *layout = new QGridLayout(this);
     layout->setContentsMargins(1, 1, 1, 1);
     layout->setSpacing(1);
 
@@ -30,69 +133,93 @@ SplitOverlay::SplitOverlay(Split *parent)
     layout->setColumnStretch(1, 1);
     layout->setColumnStretch(3, 1);
 
-    auto *move = new QPushButton(getResources().split.move, QString());
-    auto *left = this->left_ =
-        new QPushButton(getResources().split.left, QString());
-    auto *right = this->right_ =
-        new QPushButton(getResources().split.right, QString());
-    auto *up = this->up_ = new QPushButton(getResources().split.up, QString());
-    auto *down = this->down_ =
-        new QPushButton(getResources().split.down, QString());
+    this->move_ = new QPushButton(getResources().split.move, {});
+    this->left_ = new QPushButton(getResources().split.left, {});
+    this->right_ = new QPushButton(getResources().split.right, {});
+    this->up_ = new QPushButton(getResources().split.up, {});
+    this->down_ = new QPushButton(getResources().split.down, {});
 
-    move->setGraphicsEffect(new QGraphicsOpacityEffect(this));
-    left->setGraphicsEffect(new QGraphicsOpacityEffect(this));
-    right->setGraphicsEffect(new QGraphicsOpacityEffect(this));
-    up->setGraphicsEffect(new QGraphicsOpacityEffect(this));
-    down->setGraphicsEffect(new QGraphicsOpacityEffect(this));
+    this->move_->setGraphicsEffect(new QGraphicsOpacityEffect(this));
+    this->left_->setGraphicsEffect(new QGraphicsOpacityEffect(this));
+    this->right_->setGraphicsEffect(new QGraphicsOpacityEffect(this));
+    this->up_->setGraphicsEffect(new QGraphicsOpacityEffect(this));
+    this->down_->setGraphicsEffect(new QGraphicsOpacityEffect(this));
 
-    move->setFlat(true);
-    left->setFlat(true);
-    right->setFlat(true);
-    up->setFlat(true);
-    down->setFlat(true);
+    this->move_->setFlat(true);
+    this->left_->setFlat(true);
+    this->right_->setFlat(true);
+    this->up_->setFlat(true);
+    this->down_->setFlat(true);
 
-    layout->addWidget(move, 2, 2);
-    layout->addWidget(left, 2, 0);
-    layout->addWidget(right, 2, 4);
-    layout->addWidget(up, 0, 2);
-    layout->addWidget(down, 4, 2);
+    layout->addWidget(this->move_, 2, 2);
+    layout->addWidget(this->left_, 2, 0);
+    layout->addWidget(this->right_, 2, 4);
+    layout->addWidget(this->up_, 0, 2);
+    layout->addWidget(this->down_, 4, 2);
 
-    move->installEventFilter(new ButtonEventFilter(this, SplitMove));
-    left->installEventFilter(new ButtonEventFilter(this, SplitLeft));
-    right->installEventFilter(new ButtonEventFilter(this, SplitRight));
-    up->installEventFilter(new ButtonEventFilter(this, SplitUp));
-    down->installEventFilter(new ButtonEventFilter(this, SplitDown));
+    this->move_->installEventFilter(
+        new ButtonEventFilter(this, SplitOverlayButton::Move));
+    this->left_->installEventFilter(
+        new ButtonEventFilter(this, SplitOverlayButton::Left));
+    this->right_->installEventFilter(
+        new ButtonEventFilter(this, SplitOverlayButton::Right));
+    this->up_->installEventFilter(
+        new ButtonEventFilter(this, SplitOverlayButton::Up));
+    this->down_->installEventFilter(
+        new ButtonEventFilter(this, SplitOverlayButton::Down));
 
-    move->setFocusPolicy(Qt::NoFocus);
-    left->setFocusPolicy(Qt::NoFocus);
-    right->setFocusPolicy(Qt::NoFocus);
-    up->setFocusPolicy(Qt::NoFocus);
-    down->setFocusPolicy(Qt::NoFocus);
+    this->move_->setFocusPolicy(Qt::NoFocus);
+    this->left_->setFocusPolicy(Qt::NoFocus);
+    this->right_->setFocusPolicy(Qt::NoFocus);
+    this->up_->setFocusPolicy(Qt::NoFocus);
+    this->down_->setFocusPolicy(Qt::NoFocus);
 
-    move->setCursor(Qt::SizeAllCursor);
-    left->setCursor(Qt::PointingHandCursor);
-    right->setCursor(Qt::PointingHandCursor);
-    up->setCursor(Qt::PointingHandCursor);
-    down->setCursor(Qt::PointingHandCursor);
-
-    this->signalHolder_.managedConnect(this->scaleChanged, [=](float _scale) {
-        int a = int(_scale * 30);
-        QSize size(a, a);
-
-        move->setIconSize(size);
-        left->setIconSize(size);
-        right->setIconSize(size);
-        up->setIconSize(size);
-        down->setIconSize(size);
-    });
+    this->move_->setCursor(Qt::SizeAllCursor);
+    this->left_->setCursor(Qt::PointingHandCursor);
+    this->right_->setCursor(Qt::PointingHandCursor);
+    this->up_->setCursor(Qt::PointingHandCursor);
+    this->down_->setCursor(Qt::PointingHandCursor);
 
     this->setMouseTracking(true);
 
     this->setCursor(Qt::ArrowCursor);
 }
 
-void SplitOverlay::paintEvent(QPaintEvent *)
+void SplitOverlay::setHoveredButton(
+    std::optional<SplitOverlayButton> hoveredButton)
 {
+    this->hoveredButton_ = hoveredButton;
+    this->update();
+}
+
+void SplitOverlay::dragPressed()
+{
+    this->split_->drag();
+}
+
+void SplitOverlay::createSplitPressed(SplitDirection direction)
+{
+    this->split_->insertSplitRequested.invoke(direction, this->split_);
+    this->hide();
+}
+
+void SplitOverlay::scaleChangedEvent(float newScale)
+{
+    int a = int(newScale * 30);
+    QSize size(a, a);
+
+    this->move_->setIconSize(size);
+    this->left_->setIconSize(size);
+    this->right_->setIconSize(size);
+    this->up_->setIconSize(size);
+    this->down_->setIconSize(size);
+    BaseWidget::scaleChangedEvent(newScale);
+}
+
+void SplitOverlay::paintEvent(QPaintEvent *event)
+{
+    (void)event;
+
     QPainter painter(this);
     if (this->theme->isLightTheme())
     {
@@ -103,26 +230,32 @@ void SplitOverlay::paintEvent(QPaintEvent *)
         painter.fillRect(this->rect(), QColor(0, 0, 0, 150));
     }
 
-    QRect rect;
-    switch (this->hoveredElement_)
+    if (!this->hoveredButton_.has_value())
     {
-        case SplitLeft: {
+        return;
+    }
+
+    QRect rect;
+    auto hoveredButton = this->hoveredButton_.value();
+    switch (hoveredButton)
+    {
+        case SplitOverlayButton::Left: {
             rect = QRect(0, 0, this->width() / 2, this->height());
         }
         break;
 
-        case SplitRight: {
+        case SplitOverlayButton::Right: {
             rect =
                 QRect(this->width() / 2, 0, this->width() / 2, this->height());
         }
         break;
 
-        case SplitUp: {
+        case SplitOverlayButton::Up: {
             rect = QRect(0, 0, this->width(), this->height() / 2);
         }
         break;
 
-        case SplitDown: {
+        case SplitOverlayButton::Down: {
             rect =
                 QRect(0, this->height() / 2, this->width(), this->height() / 2);
         }
@@ -131,11 +264,10 @@ void SplitOverlay::paintEvent(QPaintEvent *)
         default:;
     }
 
-    rect.setRight(rect.right() - 1);
-    rect.setBottom(rect.bottom() - 1);
-
     if (!rect.isNull())
     {
+        rect.setRight(rect.right() - 1);
+        rect.setBottom(rect.bottom() - 1);
         painter.setPen(getApp()->themes->splits.dropPreviewBorder);
         painter.setBrush(getApp()->themes->splits.dropPreview);
         painter.drawRect(rect);
@@ -144,9 +276,12 @@ void SplitOverlay::paintEvent(QPaintEvent *)
 
 void SplitOverlay::resizeEvent(QResizeEvent *event)
 {
-    float _scale = this->scale();
-    bool wideEnough = event->size().width() > 150 * _scale;
-    bool highEnough = event->size().height() > 150 * _scale;
+    const auto currentScale = this->scale();
+    const auto minimumWidth = 150.F * currentScale;
+    const auto minimumHeight = 150.F * currentScale;
+
+    const auto wideEnough = event->size().width() > minimumWidth;
+    const auto highEnough = event->size().height() > minimumHeight;
 
     this->left_->setVisible(wideEnough);
     this->right_->setVisible(wideEnough);
@@ -162,77 +297,6 @@ void SplitOverlay::mouseMoveEvent(QMouseEvent *event)
     //    Qt::AltModifier) {
     //        this->hide();
     //    }
-}
-
-SplitOverlay::ButtonEventFilter::ButtonEventFilter(SplitOverlay *_parent,
-                                                   HoveredElement _element)
-    : QObject(_parent)
-    , parent(_parent)
-    , hoveredElement(_element)
-{
-}
-
-bool SplitOverlay::ButtonEventFilter::eventFilter(QObject *watched,
-                                                  QEvent *event)
-{
-    switch (event->type())
-    {
-        case QEvent::Enter: {
-            QGraphicsOpacityEffect *effect =
-                dynamic_cast<QGraphicsOpacityEffect *>(
-                    ((QWidget *)watched)->graphicsEffect());
-
-            if (effect != nullptr)
-            {
-                effect->setOpacity(0.99);
-            }
-
-            this->parent->hoveredElement_ = this->hoveredElement;
-            this->parent->update();
-        }
-        break;
-        case QEvent::Leave: {
-            QGraphicsOpacityEffect *effect =
-                dynamic_cast<QGraphicsOpacityEffect *>(
-                    ((QWidget *)watched)->graphicsEffect());
-
-            if (effect != nullptr)
-            {
-                effect->setOpacity(0.7);
-            }
-
-            this->parent->hoveredElement_ = HoveredElement::None;
-            this->parent->update();
-        }
-        break;
-        case QEvent::MouseButtonPress: {
-            if (this->hoveredElement == HoveredElement::SplitMove)
-            {
-                QMouseEvent *mouseEvent = static_cast<QMouseEvent *>(event);
-                if (mouseEvent->button() == Qt::LeftButton)
-                {
-                    this->parent->split_->drag();
-                }
-                return true;
-            }
-        }
-        break;
-        case QEvent::MouseButtonRelease: {
-            if (this->hoveredElement != HoveredElement::SplitMove)
-            {
-                auto dir = SplitContainer::Direction(
-                    this->hoveredElement + SplitContainer::Left - SplitLeft);
-
-                this->parent->split_->insertSplitRequested.invoke(
-                    static_cast<int>(dir), this->parent->split_);
-
-                this->parent->hide();
-            }
-        }
-        break;
-        default:;
-    }
-    return QObject::eventFilter(watched, event);
 }
 
 }  // namespace chatterino